Bluetooth: Update le_scan_disable_work to use HCI request
This patch updates the le_scan_disable_work function so it uses the
HCI request framework. All discovery state handling are done in the
in le_scan_disable_work_complete callback.

This patch also changes the le_scan_disable_work behavior by adding
the HCI Inquiry command to the request in case we are running the
interleaved discovery. This way we don't need the extra logic in
hci_cc_le_set_scan_enable to start the inquiry procedure.

Signed-off-by: Andre Guedes <andre.guedes@openbossa.org>