Bluetooth: Introduce the background LE scan helper
This patch adds the local function enable_background_le_scan which
starts the LE passive scanning. It will be used to enable the passive
scanning for connection establisment.

This patch also adds the hci_conn_fail_pending which does the proper
cleanup in case a connection attempt failed. It will also be used
by other functions in hci_conn.c.

Signed-off-by: Andre Guedes <andre.guedes@openbossa.org>