Bluetooth: Handle hci_conn timeout in BT_CONNECT
In order to properly cancel a LE hci_conn in BT_CONNECT state we
should check if we are scanning for devices or initiating a
connection. To identify this situation this patch adds a bool flag
in struct hci_conn to indicate if the connection is initiating. So,
if connection is initiating, we send the LE Create Connection Cancel
command. Otherwise, we remove the passive scanning trigger.

Signed-off-by: Andre Guedes <andre.guedes@openbossa.org>