Bluetooth: Change LE connection routine
In order to better support LE connection requirements, such as
multiple connections and re-connections, we should change the
current LE connection routine. Today, we use the direct
connection establishment procedure which has some limitations
and, therefore, requires extra connection management at user-space
in order to support LE connection requirements.

The new LE connection routine this patch introduces is based on
the general connection establishment procedure (described in Core
spec) and work as follows: The host starts scanning for LE devices,
once the device we want to connect to is in-range, it initiates a
connection. The procedure is terminated when the connection is
established or when the host terminates the procedure.

Signed-off-by: Andre Guedes <andre.guedes@openbossa.org>