Bluetooth: Temporarily disable passive scanning
Some LE controllers don't support scanning and initiating a connection
at the same time. So, for those controllers, we should temporarily
disable the passive scanning and enable it again once the connection
is terminated (successfully or not).

So this patch introduces the hci_conn_check_passive_scanning which
checks if the passive scanning should be enabled.

Signed-off-by: Andre Guedes <andre.guedes@openbossa.org>