Bluetooth: Postpone adding hdev to debugfs
This patch moves the hci_add_sysfs call from hci_register_dev() to
hci_dev_open(). This way hdev is only added to debugfs after it has
been properly initialized. This will enable us to export information
according to the device type (e.g. BR/EDR related information is not
exported if device is LE-only, and vice-versa).

This patch also moves the hci_del_sysfs call from hci_unregister_dev()
to hci_dev_do_close() so we keep it lined up with hci_add_sysfs call.

As a side effect, the hciX/ directory will be added to debugfs once
the adapter is powered on and it will be removed once the adapter is
powered off.

Signed-off-by: Andre Guedes <andre.guedes@openbossa.org>