Bluetooth: Add auto_connect list
This patch adds to hdev the auto_connect list. This list holds the
LE device address of remote device the kernel should automatically
connect to in case the device is in range.

This patch also adds helper functions to manipulate the auto_
connect list. Some of those functions are used in upcomming
patches.

Signed-off-by: Andre Guedes <andre.guedes@openbossa.org>