Bluetooth: Background scan and controller reset
If the controller is reset, the connections are dropped and no
Disconnection Complete Event will be sent to host. This way, no
background scan will be triggered once a auto connect device is
disconnected.

To fix this issue, before reseting the controller, we trigger the
background scan for each auto connect device we are connected to.

Signed-off-by: Andre Guedes <andre.guedes@openbossa.org>