Bluetooth: Auto connection procedure
This patch adds support for the LE auto connection procedure.

Two function are introduced to handle auto connections:

__hci_set_auto_connect(): This functions is used to set a given address
as a auto connect address.

__hci_reset_auto_connect(): This function clears hdev->auto_connect
list and removes all background scan triggers.

When an address is set as an auto connect address the background scanning
is triggered.

When we connect to an auto connect device, the background scanning
is untriggered.

If a connection with an auto connect device is terminated we trigger
the background scanning.

Signed-off-by: Andre Guedes <andre.guedes@openbossa.org>