Bluetooth: Make conn_param_get() helper non-local
This patch makes the conn_param_get() helper non-local by adding the
hci_ prefix and declaring it in hci_core.h. This helper will be used
in hci_conn.c to get the connection parameters when establishing
connections.

Since hci_conn_param_get() returns a reference to the hci_conn_params
object, it was added a refcount to struct hci_conn_params to control
its lifetime. This way, we avoid bugs such as use-after-free.

Signed-off-by: Andre Guedes <andre.guedes@openbossa.org>