Bluetooth: Use auto connection infrastructure in connect()
This patch changes connect() behavior for LE connection. Instead of
sending a HCI command right away for the controller we use the auto
connection infrastructure and let it take care of the connection
establishment.

Since BT_CONNECT represents the state where the LE Create Connection
command has been sent to the controller, we should not set it in hci_
connect_le() anymore. We leave hci_conn state in BT_OPEN. This patch
also changes hci_conn_timeout() so it correctly handles connection
timeout.

As a side effect, the kernel is now able to handle multiple connect()
calls from user-space.

Signed-off-by: Andre Guedes <andre.guedes@openbossa.org>