ether: Add ETH_P_AVTP macro
This patch adds the ETH_P_AVTP macro which defines the Audio/Video
Transport Protocol (AVTP) ethertype assigned to 0x22F0, according to:

http://standards-oui.ieee.org/ethertype/eth.txt

AVTP is the transport protocol utilized in Audio/Video Bridging (AVB),
and it is defined by IEEE 1722 standard.

Note that we have ETH_P_TSN macro defined with the number assigned to
AVTP. However, there is no "TSN" ethertype. TSN is not a protocol, but a
set of features to deliver networking determinism, so ETH_P_TSN can be a
bit misleading. For compatibility reasons we should keep it around.
This patch re-defines it using the ETH_P_AVTP macro to make it explicit.

Signed-off-by: Andre Guedes <andre.guedes@intel.com>