e1000e: fix S0ix flows for cable connected case
Added a fix to S0ix entry and exit flows for TGP and above
MAC types, to the case when the ethernet cable is connected
and the link is up. With that the system is able to reach
SLP_S0 when going to freeze power state.

Change-type: ImplementationChange
Title: e1000e: fix S0ix flows for cable connected case
Signed-off-by: Vitaly Lifshits <vitaly.lifshits@intel.com>