intel-ethernet: make W=1 build cleanly
This takes care of all of the trivial W=1 fixes in the Intel
Ethernet drivers, which allows developers and maintainers to
build more of the networking tree with more complete warning
checks.

Almost all of the changes were trivial comment updates on
function headers, but some of the changes were for variables that
were storing a return value from a register read, where the
return value wasn't used. Those conversions to remove the lvalue
of the assignment should be safe because the readl memory mapped
reads are marked volatile and should not be optimized out without
an lvalue (I suspect a very long time ago this wasn't guaranteed
as it is today).

Inspired by Lee Jones' series of wireless work to do the same.
Compile tested only.

Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>